Cambios proyecto final

+++ ProyectoFinal/agenciainmobiliaria/src/main/java/co/com/udem/agenciainmobiliaria/filtros/FilterSpecification.java
@@ -1,5 +1,7 @@
 package co.com.udem.agenciainmobiliaria.filtros;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.domain.Specification;
 
 import co.com.udem.agenciainmobiliaria.entities.Propiedad;
@@ -11,10 +13,10 @@ public class FilterSpecification {
 	public static Specification<Propiedad> withFilter(Object object, String columName) {
 		return object == null ? null : (root, query, cb) -> cb.equal(root.get(columName), object);
 	}
-	
-	
-	 public static Specification<Propiedad> withFilterBetween(Double precioIni, Double precioFinal, String columName) {
-	        return (precioIni == null || precioFinal == null) ? null
-	                : (root, query, cb) -> cb.between(root.get(columName), precioIni, precioFinal);
-	    }
+
+	public static Specification<Propiedad> withFilterBetween(BigInteger precioIni, BigInteger precioFinal,
+			String columName) {
+		return (precioIni == null || precioFinal == null) ? null
+				: (root, query, cb) -> cb.between(root.get(columName), precioIni, precioFinal);
+	}
 }
